still refactoring and still working
