+++ tests/models/test_message.py
@@ -10,6 +10,100 @@ from google.cloud.pubsub_v1 import PublisherClient
 from models.message import File, Message, create_message, send_pub_sub_message
 
 
+def test_file_extension(file):
+    assert file.extension() == ".zip"
+
+
+def test_file_filename(file):
+    assert file.filename() == "dd_file.zip"
+
+
+@pytest.mark.parametrize(
+    "file_name,file_type",
+    [("dd_file.zip", "dd"), ("mi_file.zip", "mi")],
+)
+def test_file_type(file, file_name, file_type):
+    file.name = f"{file_name}:my-bucket-name"
+    assert file.type() == file_type
+
+
+@pytest.mark.parametrize(
+    "file_name, expected",
+    [
+        ("dd_opn2101a.zip", "OPN"),
+        ("dd_lms2102_a1.zip", "LMS"),
+        ("dd_lms2102_bk1.zip", "LMS"),
+        ("dd_lmc2102_bk1.zip", "LMC"),
+        ("dd_lmb21021_bk2.zip", "LMB"),
+    ],
+)
+def test_file_survey_name(file, file_name, expected):
+    file.name = f"{file_name}:my-bucket-name"
+    assert file.survey_name() == expected
+
+
+@pytest.mark.parametrize(
+    "file_name, expected",
+    [
+        ("dd_opn2101a.zip", "OPN2101A"),
+        ("dd_lms2102_a1.zip", "LMS2102_A1"),
+        ("dd_lms2102_bk1.zip", "LMS2102_BK1"),
+        ("dd_lmc2102_bk1.zip", "LMC2102_BK1"),
+    ],
+)
+def test_file_instrument_name(file, file_name, expected):
+    file.name = f"{file_name}:my-bucket-name"
+    assert file.instrument_name() == expected
+
+
+@pytest.mark.parametrize(
+    "survey_name, expected",
+    [
+        ("OPN", False),
+        ("OLS", False),
+        ("LMS", True),
+        ("LMB", True),
+        ("IPS", False),
+        ("LMC", True),
+        ("LMO", True),
+        ("QWERTY", False),
+        ("LMNOP", True),
+        ("LBS", False),
+    ],
+)
+def test_file_is_lms(file, survey_name, expected):
+    file.name = f"dd_{survey_name}2101a.zip:my-bucket-name"
+    assert file.is_lms() is expected
+
+
+@pytest.mark.parametrize(
+    "survey_name, expected",
+    [
+        ("OPN", True),
+        ("opn", True),
+        ("LMS", False),
+        ("LMB", False),
+        ("IPS", False),
+        ("LMC", False),
+        ("LMO", False),
+        ("QWERTY", False),
+        ("LMNOP", False),
+        ("OPNFOOBAR", True),
+    ],
+)
+def test_file_is_opn(file, survey_name, expected):
+    file.name = f"dd_{survey_name}2101a.zip:my-bucket-name"
+    assert file.is_opn() is expected
+
+
+def test_file_from_event(dd_event):
+    file = File.from_event(dd_event("OPN2102R"))
+    assert file.name == "dd_OPN2102R_0103202021_16428.zip:ons-blaise-v2-nifi"
+    assert file.sizeBytes == "20"
+    assert file.md5sum == "d1ad7875be9ee3c6fde3b6f9efdf3c6b67fad78ebd7f6dbc"
+    assert file.relativePath == ".\\"
+
+
 def test_message_json():
     file = File(name="a_file", sizeBytes="20", md5sum="dasdasd", relativePath="./")
     message = Message(
@@ -44,98 +138,8 @@ def test_message_json():
     }
 
 
-def test_file_extension(file):
-    assert file.extension() == ".zip"
-
-
-def test_file_filename(file):
-    assert file.filename() == "dd_file.zip"
-
-
-@pytest.mark.parametrize(
-    "file_name,file_type",
-    [("dd_file.zip", "dd"), ("mi_file.zip", "mi")],
-)
-def test_file_file_type(file, file_name, file_type):
-    file.name = f"{file_name}:my-bucket-name"
-    assert file.type() == file_type
-
-
-@pytest.mark.parametrize(
-    "file_name, expected",
-    [
-        ("dd_opn2101a.zip", "OPN"),
-        ("dd_lms2102_a1.zip", "LMS"),
-        ("dd_lms2102_bk1.zip", "LMS"),
-        ("dd_lmc2102_bk1.zip", "LMC"),
-        ("dd_lmb21021_bk2.zip", "LMB"),
-    ],
-)
-def test_file_survey_name(file, file_name, expected):
-    file.name = f"{file_name}:my-bucket-name"
-    assert file.survey_name() == expected
-
-
-@pytest.mark.parametrize(
-    "file_name, expected",
-    [
-        ("dd_opn2101a.zip", "OPN2101A"),
-        ("dd_lms2102_a1.zip", "LMS2102_A1"),
-        ("dd_lms2102_bk1.zip", "LMS2102_BK1"),
-        ("dd_lmc2102_bk1.zip", "LMC2102_BK1"),
-    ],
-)
-def test_file_instrument_name(file, file_name, expected):
-    file.name = f"{file_name}:my-bucket-name"
-    assert file.instrument_name() == expected
-
-
-def test_file_from_event(dd_event):
-    file = File.from_event(dd_event("OPN2102R"))
-    assert file.name == "dd_OPN2102R_0103202021_16428.zip:ons-blaise-v2-nifi"
-    assert file.sizeBytes == "20"
-    assert file.md5sum == "d1ad7875be9ee3c6fde3b6f9efdf3c6b67fad78ebd7f6dbc"
-    assert file.relativePath == ".\\"
-
-
-@pytest.mark.parametrize(
-    "survey_name, expected",
-    [
-        ("OPN", False),
-        ("OLS", False),
-        ("LMS", True),
-        ("LMB", True),
-        ("IPS", False),
-        ("LMC", True),
-        ("LMO", True),
-        ("QWERTY", False),
-        ("LMNOP", True),
-        ("LBS", False),
-    ],
-)
-def test_is_lms(file, survey_name, expected):
-    file.name = f"dd_{survey_name}2101a.zip:my-bucket-name"
-    assert file.is_lms() is expected
-
-
-@pytest.mark.parametrize(
-    "survey_name, expected",
-    [
-        ("OPN", True),
-        ("opn", True),
-        ("LMS", False),
-        ("LMB", False),
-        ("IPS", False),
-        ("LMC", False),
-        ("LMO", False),
-        ("QWERTY", False),
-        ("LMNOP", False),
-        ("OPNFOOBAR", True),
-    ],
-)
-def test_is_opn(file, survey_name, expected):
-    file.name = f"dd_{survey_name}2101a.zip:my-bucket-name"
-    assert file.is_opn() is expected
+def test_message_first_file(file):
+    foo = file
 
 
 def test_create_message_mi(mi_event, config):
